강아지 정보 수정 오류 해결
1. 강아지 이름 변경 할 때 오류 메시지 뜨는 거 해결 -> 강아지 사진이 있는 경우, 강아지 이름을 변경 할 때, 이전의 사진을 지우고 새로운 이름의 사진을 생성해야함. 하지만 사진이 없는 경우에는 강아지 이름을 변경 할 때, 이전의 사진이 없으므로 없는 사진을 지우는 오류가 생김. 그래서 이를 이전의 사진이 있는 지 판단하는 코드를 추가하여 해결.
2. 산책 저장에 관한 뷰모델, 유저 정보 수정에 대한 뷰모델, 강아지 정보 수정에 대한 뷰모델 구현.

+++ app/src/main/java/com/example/walkingdogapp/repository/UserInfoRepository.kt
@@ -428,6 +428,7 @@ class UserInfoRepository @Inject constructor(
                         }
                     }
                 } catch (e: Exception) {
+                    Log.d("savepoint1", e.message.toString())
                     error = true
                     return@upload
                 }
@@ -441,7 +442,7 @@ class UserInfoRepository @Inject constructor(
                         return@delete
                     }
 
-                    if (!MainActivity.dogNameList.contains(dogInfo.name)) {
+                    if (!MainActivity.dogNameList.contains(dogInfo.name) && MainActivity.dogUriList[beforeName] != null) {
                         storageRef.child(beforeName).delete().await()
                     }
                 } catch (e: Exception) {
